Back to spaceship fixes
